LayoutTests/fast/dom/Range/clone-contents-document-type.html

<!DOCTYPE>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Checks that Range.cloneContents() throws a HierarchyRequestError when encountering a docType");

var range = document.createRange();
range.setStart(document, 0);
range.setEnd(document, 2);

shouldThrow("range.cloneContents()", "'Error: HierarchyRequestError: DOM Exception 3'");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
