LayoutTests/fast/dom/Range/clone-contents-document-type-expected.txt

Checks that Range.cloneContents() throws a HierarchyRequestError when encountering a docType

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range.cloneContents() threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS successfullyParsed is true

TEST COMPLETE
